Set the notifications webhook by URL to prevent false alerts (#952)
In the event that someone who has forked this repository sets up CircleCI integrations for it, we would like that builds on their master branch would not be able to send alerts to the Hyperledger Rocket Chat #besu-contributors channel. Here we move to having the notification web-hook be sent to an address set by the HUBOT_URL variable instead. This environment variable is set in the CircleCI web-app in the same place we set the variables for our secrets for other 3rd party app access.

It has been tested that an environment without the variable being set would cause no issues for the CircleCI build system either for other forked projects.

Signed-off-by: Edward Evans <edward.joshua.evans@gmail.com>
Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>