graalvm
WIP to test out speed.

It's a meh.  bench-static run against a graalvm toy and evmtool shows
that hotspot outperforms graal once its warmed up (100 iterations per op
or so)  cold start is much better and may be useful for test generation,
but for long running processes (more than a minute) graal is not worth
the time to fully implement.

Signed-off-by: Danno Ferrin <danno.ferrin@swirldslabs.com>